feat: module added for three methods
