+++ lib/customer.rb
@@ -14,7 +14,7 @@ class Customer
   end
 
   def update(attributes)
-    @id = attributes[:id] unless attributes[:id].nil?
+    # @id = attributes[:id] unless attributes[:id].nil?
     @first_name = attributes[:first_name] unless attributes[:first_name].nil?
     @last_name = attributes[:last_name] unless attributes[:last_name].nil?
     @updated_at = Time.now
