+++ lib/customer_repository.rb
@@ -1,7 +1,9 @@
 require 'csv'
 require_relative 'customer'
+require_relative 'methodable'
 
 class CustomerRepository
+  include Methodable
   attr_reader :all
 
   def initialize(filepath)
@@ -17,31 +19,30 @@ class CustomerRepository
   end
 
   def find_all_by_first_name(first_name)
-    @all.find_all { |customer| customer.first_name == first_name }
+    @all.find_all { |customer| customer.first_name == first_name}
   end
 
   def find_all_by_last_name(last_name)
-    @all.find_all { |customer| customer.last_name == last_name }
+    @all.find_all { |customer| customer.last_name == last_name}
   end
 
   def create(attributes)
-    # require "pry"; binding.pry
     attributes[:id] = @all.max_by { |customer| customer.id }.id + 1
     @all << Customer.new(attributes)
     @all.last
   end
 
-  def update(id, attributes)
-    if find_by_id(id)
-      find_by_id(id).update(attributes)
-    end
-  end
+  # def update(id, attributes)
+  #   if find_by_id(id)
+  #     find_by_id(id).update(attributes)
+  #   end
+  # end
 
   def delete(id)
     @all.delete_if { |customer| customer.id == id }
   end
 
-  def inspect
-    "#<#{self.class} #{@merchants.size} rows>"
-  end
+  # def inspect
+  #   "#<#{self.class} #{@merchants.size} rows>"
+  # end
 end
